move ApprovalStates from Deployment to types/interfaces and update all imports
