+++ server/routes/v1/requests.ts
@@ -1,7 +1,8 @@
 import bodyParser from 'body-parser'
 import { Request, Response } from 'express'
 import { Types } from 'mongoose'
-import { ApprovalStates, DeploymentDoc } from '../../models/Deployment'
+import { DeploymentDoc } from '../../models/Deployment'
+import { ApprovalStates } from '../../../types/interfaces'
 import { ModelDoc } from '../../models/Model'
 import { RequestTypes } from '../../models/Request'
 import { VersionDoc } from '../../models/Version'
